Builds linux32 and linux64
idk how to debug this on linux, im using wsl

+++ ivp/ivp_collision/ivp_compact_ledge_solver.cxx
@@ -698,7 +698,7 @@ IVP_BOOL IVP_Compact_Ledge_Solver::check_ledge(const IVP_Compact_Ledge *cl)
     // For debugging only
     // Checks for consistency
 
-	IVP_ASSERT ( (int(cl) & 15) == 0);
+	IVP_ASSERT ( (size_t(cl) & 15) == 0);
 	const IVP_Compact_Poly_Point *ppppp=cl->get_point_array();
 	IVP_ASSERT ( ((size_t)ppppp & 15 ) == 0);
 
@@ -707,7 +707,7 @@ IVP_BOOL IVP_Compact_Ledge_Solver::check_ledge(const IVP_Compact_Ledge *cl)
     for(int i=0; i<cl->get_n_triangles(); i++){
 	IVP_ASSERT(i == tri->get_tri_index());
 
-		IVP_ASSERT ( (int(tri) & 15) == 0);
+		IVP_ASSERT ( (size_t(tri) & 15) == 0);
 
 	// all edges
 	for(int j=0; j<3; j++){
