+++ vphysics/main.cpp
@@ -176,7 +176,7 @@ void CPhysicsInterface::DestroyObjectPairHash( IPhysicsObjectPairHash *pHash )
 // holds a cache of these by id.
 // NOTE: This is stuffed into vphysics.dll as a sneaky way of sharing the memory between
 // client and server in single player.  So you can't have different client/server rules.
-IPhysicsCollisionSet *CPhysicsInterface::FindOrCreateCollisionSet( unsigned int id, int maxElementCount )
+IPhysicsCollisionSet *CPhysicsInterface::FindOrCreateCollisionSet( uintptr_t id, int maxElementCount )
 {
 	if ( !m_pCollisionSetHash )
 	{
@@ -190,7 +190,7 @@ IPhysicsCollisionSet *CPhysicsInterface::FindOrCreateCollisionSet( unsigned int 
 	IPhysicsCollisionSet *pSet = FindCollisionSet( id );
 	if ( pSet )
 		return pSet;
-	int index = m_collisionSets.AddToTail();
+	intptr_t index = m_collisionSets.AddToTail();
 	m_pCollisionSetHash->add_elem( (void *)id, (void *)(index+1) );
 	return &m_collisionSets[index];
 }
